Fixed a few bugs in Parser.java
We were using matches() instead of isKind() to check parentheses. matches() expects two arguments and we were only passing one.  We only need to check the kind for parentheses, so isKind is the proper method to use.

Fixed a misnamed Node. It was named first when it should have been second.

Fixed use of token when type was not declared.